[hardware] :bug: Lanes accept and grant their incoming masks
The introduction of the spill register at the edge of the lanes
to cut the critical path in2out of mask_ready_o introduces an
issue: the lanes now acknowledge and store all the incoming
mask requests, even if they are for other units (e.g., sldu).
This brings two problems: the spill reg saves a mask value
that cannot be drained since the request was not for the lanes,
and the lanes acknowledge the mask unit even if the correct
unit destination of the request (e.g., the sldu) was not ready
and did not registered the mask.
Now the ALU and the MFPU tell the spill register if they are
waiting for a request from the mask unit: if not, the spill
the incoming mask_valid is masked and the spill register
is does not acknoledge/store the incoming mask.