[hardware] :bug: Fix opreq vl for vslidedown
The number of elements to be requested for a vslidedown was wrong.
We ask full words (8*NrLanes Byte) to the VRF.
If the stride does not divide a VRF word, we will have an offset
that ranges from (0 -> 255) B. We will need to ask all these
additional Bytes as well.
If a full word does not divide vl, we will also ask for all
the additional Bytes, from (0 -> 255) B.
Remember that here the vl refers to the elements to be requested
within a single Lane, so we must also divide by NrLanes, and ask
an additional element if NrLanes does not divide vl.