[hardware] :bug: Reshuffle source vregs for in-lane ops
In case of an operation executed within the lanes (e.g., vadd),
the lanes cannot read the vector correctly if the vsew of the
source register is different from the eew_q of that register
(different byte layout!). So, we need to reshuffle in this case
as well. Since this is costly, the suggested software procedure
is to avoid filling a register with an encoding and using it
with a different one.