[hardware] Refactor STU-exception flush engine
When a store-translation exception occurs, Ara's backend
should be flushed as soon as the instruction preceding the
store is over. This happens when ara is idle again.
Ara's frontend (dispatcher, sequencer) are already "flushed".
Therefore, we wait for idle, and then we inject a flush
command to flush the pipeline following the data stream to be
sure not to miss any data around. We don't flush all at the same time
to reduce the timing pressure. Now, we have a non-timing-critical
long chain of FFs. Once the flush come back from the store unit,
the process is over.