[hardware] :bug: Kill main seq request upon stu MMU exception
In case of an exception on this burst, kill the request.
Exceptions on this burst mean that all the valid sources
have been fetched from VRF already. Don't immediately
kill when detecting the exception in the addrgen, as previous
valid bursts can still need operands to be fetched from the VRF.